Update project details and refactor models
Updated the project details in pom.xml from 'Ticket Master' to 'Learn to Code'. Refactored the models, Reply and Question. In Reply model, removed unused imports and corrected a typo in an annotation message. In Question model, added fields for imageUrls and categories. Also moved the ReplyRepository to its appropriate package from the models package.