Replace EventSerializer with ReplyMapper and refactor ReplyDto
EventSerializer has been deleted and a new ReplyMapper for reply entities has been added. This maps Reply to ReplyDto and vice versa for proper serialization and deserialization. Also, the ReplyDto class was refactored from @Value to @Data, indicating that fields can change, and it no longer implements Serializable. A field in Reply model was renamed from 'replyText' to 'reply'.